MG-396 refactoring folders
