+++ README.md
@@ -29,7 +29,7 @@
 3. That's it :)
 
 
-# Building from source (You probably don't want that)
+# Building from source (You probably don't need to do that)
 
 	$ docker-compose up prod-api-build-bk && docker-compose build prod-build-bk
